Add FingerprintWithInfo method
Introduces a new FingerprintWithInfo which returns a map where the keys
contain the names of the identified technologies and their associated
values contain basic information about the technology/App which is, for
now, description and website.

In order for this feature to work, the update-fingerprints command has
been modified to export description and website information for each
fingerprint.